Refactor components and imports
