+++ src/components/Article/index.ts
@@ -1,0 +1,5 @@
+export { ArticleDetail } from "./ArticleDetail";
+export { ArticleDetailLoading } from "./ArticleDetailLoading";
+export { ArticleDetailError } from "./ArticleDetailError";
+export { ArticleDetailNotFound } from "./ArticleDetailNotFound";
+export { ArticleDetailContent } from "./ArticleDetailContent";
