+++ src/components/ArticleCard.tsx
@@ -1,6 +1,6 @@
 import { memo } from "react";
 import { Link } from "react-router-dom";
-import type { Article } from "../types/article";
+import type { Article } from "@/types/article";
 
 interface ArticleCardProps {
   article: Article;
